+++ include/RoomData.h
@@ -52,6 +52,7 @@ class StaticModel {
   public:
     StaticModel(glm::vec3 pos, float angle, int i);
     void display(glm::mat4 VP);
+    void displayUI();
 
   private:
     int id;
@@ -76,12 +77,17 @@ class RoomSprite {
 class Portal {
   public:
     Portal(int adj, glm::vec3 n, glm::vec3 v1, glm::vec3 v2, glm::vec3 v3,
-           glm::vec3 v4) : adjoiningRoom(adj), normal(n) {
+           glm::vec3 v4) : adjoiningRoom(adj), normal(n), bbox(v1, v3) {
         vert[0] = v1; vert[1] = v2;
         vert[2] = v3; vert[3] = v4;
     }
+
     int getAdjoiningRoom() { return adjoiningRoom; }
     glm::vec3 getNormal() { return normal; }
+    BoundingBox getBoundingBox() { return bbox; }
+
+    void display(glm::mat4 VP);
+    void displayUI();
 
     glm::vec3 getVertex(int i) {
         orAssertGreaterThanEqual(i, 0);
@@ -89,10 +95,16 @@ class Portal {
         return vert[i];
     }
 
+    static void setShowBoundingBox(bool s) { showBoundingBox = s; }
+    static bool getShowBoundingBox() { return showBoundingBox; }
+
   private:
     int adjoiningRoom;
     glm::vec3 normal;
     glm::vec3 vert[4];
+    BoundingBox bbox;
+
+    static bool showBoundingBox;
 };
 
 // --------------------------------------