+++ include/StaticMesh.h
@@ -17,6 +17,7 @@ class StaticMesh {
     StaticMesh(int i, int m, BoundingBox* b1, BoundingBox* b2)
         : id(i), mesh(m), bbox1(b1), bbox2(b2) { }
     void display(glm::mat4 MVP);
+    void displayUI();
 
     int getID() { return id; }