+++ src/StaticMesh.cpp
@@ -9,14 +9,23 @@
 #include "World.h"
 #include "StaticMesh.h"
 
+#include "imgui/imgui.h"
+
 bool StaticMesh::showBoundingBox = false;
 
 void StaticMesh::display(glm::mat4 MVP) {
-    getWorld().getMesh(mesh).display(MVP);
+    World::getMesh(mesh).display(MVP);
 
     if (showBoundingBox) {
         bbox1->display(MVP, glm::vec3(1.0f, 0.0f, 0.0f), glm::vec3(0.0f, 1.0f, 0.0f));
         bbox2->display(MVP, glm::vec3(1.0f, 0.0f, 1.0f), glm::vec3(0.0f, 0.0f, 1.0f));
     }
+}
+
+void StaticMesh::displayUI() {
+    ImGui::Text("%03d", id);
+    ImGui::NextColumn();
+    ImGui::Text("%03d // and two bboxes", mesh);
+    ImGui::NextColumn();
 }